requirements.txt
Django==2.0.3
django-bootstrap3
pylint-django
django-braces
